+++ zetaclient/config/config.go
@@ -4,16 +4,18 @@ package config
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
 	"sync"
 
 	"github.com/asaskevich/govalidator"
+	sdktypes "github.com/cosmos/cosmos-sdk/types"
 	"github.com/fsnotify/fsnotify"
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
+
+	"github.com/zeta-chain/node/pkg/chains"
 )
 
 // restrictedAddressBook is a map of restricted addresses
@@ -32,7 +34,7 @@ const folder string = "config"
 func Save(config *Config, path string) error {
 	// validate config
 	if err := Validate(*config); err != nil {
-		return err
+		return errors.Wrapf(err, "config file validation failed")
 	}
 
 	folderPath := filepath.Join(path, folder)
@@ -90,27 +92,64 @@ func Load(basePath string) (Config, error) {
 
 	// validate config
 	if err := Validate(cfg); err != nil {
-		return Config{}, err
+		return Config{}, errors.Wrapf(err, "config file validation failed")
 	}
 
 	return cfg, nil
 }
 
 // Validate performs basic validation on the config fields
-// TODO: add more validation for other fields
-// https://github.com/zeta-chain/node/issues/4352
 func Validate(cfg Config) error {
 	// go-tss requires a valid IPv4 address
 	if cfg.PublicIP != "" && !govalidator.IsIPv4(cfg.PublicIP) {
-		return fmt.Errorf("invalid public IP %s", cfg.PublicIP)
+		return errors.Errorf("reason: invalid public IP, got: %s", cfg.PublicIP)
 	}
 
 	if cfg.PublicDNS != "" && !govalidator.IsDNSName(cfg.PublicDNS) {
-		return fmt.Errorf("invalid public DNS %s", cfg.PublicDNS)
+		return errors.Errorf("reason: invalid public DNS, got: %s", cfg.PublicDNS)
+	}
+
+	if _, err := chains.ZetaChainFromCosmosChainID(cfg.ChainID); err != nil {
+		return errors.Errorf("reason: invalid chain id, got: %s", cfg.ChainID)
+	}
+
+	// ZetaCoreURL can be either an IP address or a hostname (e.g., Docker service name)
+	if cfg.ZetaCoreURL != "" && !govalidator.IsIP(cfg.ZetaCoreURL) && !govalidator.IsDNSName(cfg.ZetaCoreURL) {
+		return errors.Errorf("reason: invalid zetacore URL, got: %s", cfg.ZetaCoreURL)
+	}
+
+	// validate granter address - should be a valid bech32 address
+	if _, err := sdktypes.AccAddressFromBech32(cfg.AuthzGranter); err != nil {
+		return errors.Errorf("reason: invalid bech32 granter address, got: %s", cfg.AuthzGranter)
+	}
+
+	// validate grantee name - should not be empty
+	if strings.TrimSpace(cfg.AuthzHotkey) == "" {
+		return errors.Errorf("reason: grantee name is empty")
+	}
+
+	// acceptable log levels are: 0:debug, 1:info, 2:warn, 3:error, 4:fatal, 5:panic
+	if cfg.LogLevel < 0 || cfg.LogLevel > 5 {
+		return errors.Errorf("reason: log level must be between 0 and 5, got: %d", cfg.LogLevel)
+	}
+
+	if cfg.ConfigUpdateTicker == 0 {
+		return errors.Errorf("reason: config update ticker is 0")
 	}
 
 	if cfg.KeyringBackend != KeyringBackendFile && cfg.KeyringBackend != KeyringBackendTest {
-		return fmt.Errorf("invalid keyring backend %s", cfg.KeyringBackend)
+		return errors.Errorf("reason: invalid keyring backend, got: %s", cfg.KeyringBackend)
+	}
+
+	if cfg.MaxBaseFee < 0 {
+		return errors.Errorf("reason: max base fee cannot be negative, got: %d", cfg.MaxBaseFee)
+	}
+
+	if cfg.MempoolCongestionThreshold < 0 {
+		return errors.Errorf(
+			"reason: mempool congestion threshold cannot be negative, got: %d",
+			cfg.MempoolCongestionThreshold,
+		)
 	}
 
 	return nil